update first and re-launch
not background

fix sleep

test

should fetch after ui prompt

test

rm

+++ launch_openpilot.sh
@@ -1,32 +1,38 @@
 #!/usr/bin/bash
 # wait for weston to come up
 sleep 5
+pkill -f spinner
 
 if [ -z "$BASEDIR" ]; then
   BASEDIR="/data/openpilot"
 fi
 export PYTHONPATH="/data/openpilot"
 
-pkill -f spinner
-# launch spinner
+# build
 echo "Building" | ./selfdrive/ui/spinner &
 spinner_pid=$!
 
-# build
 scons selfdrive/ui/_notouch selfdrive/ui/_ui -j8
 kill -9 $spinner_pid
 
 # launch ui and let user set up ssh
 ./selfdrive/ui/ui
 
-echo "Fetching updates" | ./selfdrive/ui/spinner &
-spinner_pid=$!
+# update if not done already
+if [[ "$UPDATE_DONE" != "true" ]]; then
+  echo "Fetching updates" | ./selfdrive/ui/spinner &
+  spinner_pid=$!
 
-# update
-git fetch origin $(git rev-parse --abbrev-ref HEAD)
-git reset --hard "@{u}"
-git submodule update --init --recursive -f
-kill -9 $spinner_pid
+  git fetch origin $(git rev-parse --abbrev-ref HEAD)
+  git reset --hard "@{u}"
+  git submodule update --init --recursive -f
+  kill -9 $spinner_pid
+
+  # export update status so next run doesn't fetch
+  export UPDATE_DONE="true"
+  exec /bin/bash "$0"
+  exit 0
+fi
 
 # build again after update
 echo "Building" | ./selfdrive/ui/spinner &
